Fixed test; some refactoring; added some more comments
